EnvironmentPlugin: Accept empty env strings
It's valid to have an empty string as an environment variable, and it
shouldn't be treated the same way as an undefined variable.

The previous code, designed to fall back to a default, would basically
evaluate to `"" || undefined` which evaluates to `undefined`. This would
then cause the warning that the environment variable was undefined.

Instead, we should use the empty string value and not fall back to the
default.